requirements.txt
attrs==19.3.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.2
Flask-WTF==0.14.3
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
lxml==4.5.0
MarkupSafe==1.1.1
newspaper3k==0.2.8
nltk==3.5
Pillow==7.1.1
pyrsistent==0.16.0
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
requests-file==1.4.3
six==1.14.0
soupsieve==2.0
tinysegmenter==0.3
tldextract==2.2.2
tqdm==4.45.0
urllib3==1.25.9
validator-collection==1.4.1
Werkzeug==1.0.1
WTForms==2.2.1
zipp==3.1.0
